refactor: repository 클래스가 dto 패키지에 있던 부분 수정, 컨트롤러 url 수정
