+++ OurMemory_Server/src/main/java/com/kds/ourmemory/controller/v1/user/UserController.java
@@ -2,10 +2,10 @@ package com.kds.ourmemory.controller.v1.user;
 
 import java.util.Date;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -16,16 +16,15 @@ import com.kds.ourmemory.dto.user.SignUpResponseDto;
 import com.kds.ourmemory.service.v1.user.UserService;
 
 import io.swagger.annotations.ApiOperation;
+import lombok.RequiredArgsConstructor;
 
-@RestController(value = "/v1")
+@RequiredArgsConstructor
+@RestController
+@RequestMapping(value = "/v1")
 public class UserController {
-	private UserService service;
+    
+	private final UserService service;
 	
-	@Autowired
-	public UserController(UserService service) {
-		this.service= service;
-	}
-
 	@ApiOperation(value="회원가입", notes = "앱에서 전달받은 데이터로 회원가입 진행")
 	@PostMapping("/signUp")
 	public SignUpResponseDto signUp(@RequestBody SignUpRequestDto request) {
