+++ OurMemory_Server/src/main/java/com/kds/ourmemory/domain/UsersAndMemorysPk.java
@@ -3,18 +3,20 @@ package com.kds.ourmemory.domain;
 import java.io.Serializable;
 
 import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+@EqualsAndHashCode
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class UsersAndMemorysPk implements Serializable{
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	
-	private Users users;
-	private Memorys memorys;
+class UsersAndMemorysPk implements Serializable{
+    /**
+     * 
+     */
+    private static final long serialVersionUID = 1L;
+    
+    private Users users;
+    private Memorys memorys;
 }
